requirements.txt
Flask~=3.1.2
python-dotenv~=1.1.0
requests~=2.32.3
pytest~=8.3.5
idna~=3.10
httpcore~=1.0.9
httpx~=0.28.1
anyio~=4.9.0
h11~=0.16.0
pip~=23.2.1
typing_extensions~=4.13.2
packaging~=25.0
pydantic~=2.11.3
click~=8.3.0
colorama~=0.4.6
pydantic_core~=2.33.1
pathspec~=0.12.1
urllib3~=2.3.0
cachetools~=5.5.2
Jinja2~=3.1.6
sniffio~=1.3.1
charset-normalizer~=3.4.1
platformdirs~=4.4.0
multipledispatch~=1.0.0
MarkupSafe~=3.0.2
websockets~=15.0.1
certifi~=2025.1.31
annotated-types~=0.7.0
itsdangerous~=2.2.0
blinker~=1.9.0
Werkzeug~=3.1.3
path~=17.1.1
beautifulsoup4~=4.13.4